+++ ee/LICENSE
@@ -1,0 +1,1 @@
+The BoxyHQ Enterprise Edition (EE) license (the “EE License”)
