+++ ee/federated-saml/pages/edit.tsx
@@ -1,51 +1,78 @@
 import type { NextPage } from 'next';
 import type { SAMLFederationApp } from '@boxyhq/saml-jackson';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import Link from 'next/link';
 import toast from 'react-hot-toast';
 import classNames from 'classnames';
 import { useRouter } from 'next/router';
 import { useTranslation } from 'next-i18next';
 import { ArrowLeftIcon } from '@heroicons/react/24/outline';
+import useSWR from 'swr';
 
-import { ApiResponse } from 'types';
+import { fetcher } from '@lib/ui/utils';
+import { ApiSuccess, ApiError, ApiResponse } from 'types';
+import Loading from '@components/Loading';
+import Alert from '@components/Alert';
 
 const initialValue = {
   name: '',
-  tenant: '',
-  product: '',
   acsUrl: '',
   entityId: '',
 };
 
-const NewApp: NextPage = () => {
+const UpdateApp: NextPage = () => {
   const { t } = useTranslation('common');
   const router = useRouter();
   const [loading, setLoading] = useState(false);
-  const [newApp, setApp] = useState(initialValue);
+  const [existingApp, setApp] = useState<typeof initialValue>(initialValue);
+
+  const { id } = router.query as { id: string };
+
+  const { data, error } = useSWR<ApiSuccess<SAMLFederationApp>, ApiError>(
+    `/api/admin/federated-saml/${id}`,
+    fetcher,
+    {
+      revalidateOnFocus: false,
+    }
+  );
+
+  useEffect(() => {
+    if (data) {
+      setApp(data.data);
+    }
+  }, [data]);
+
+  if (error) {
+    return <Alert type='error' message={error.message}></Alert>;
+  }
+
+  if (!data) {
+    return <Loading />;
+  }
+
+  const app = data.data;
 
   const onSubmit = async (event: React.FormEvent) => {
     event.preventDefault();
 
     setLoading(true);
 
-    const response = await fetch('/api/admin/federated-saml', {
-      method: 'POST',
+    const response = await fetch(`/api/admin/federated-saml/${app.id}`, {
+      method: 'PUT',
       headers: {
         'Content-Type': 'application/json',
       },
-      body: JSON.stringify(newApp),
+      body: JSON.stringify(existingApp),
     });
 
     setLoading(false);
 
-    const { data: app, error }: ApiResponse<SAMLFederationApp> = await response.json();
+    const { error }: ApiResponse<SAMLFederationApp> = await response.json();
 
     if (!response.ok) {
       toast.error(error.message);
     } else {
-      router.replace(`/admin/federated-saml/${app.id}/metadata`);
-      toast.success('SAML federation app created successfully');
+      toast.success('SAML federation app updated successfully');
     }
   };
 
@@ -53,7 +80,7 @@ const NewApp: NextPage = () => {
     const target = event.target as HTMLInputElement;
 
     setApp({
-      ...newApp,
+      ...existingApp,
       [target.id]: target.value,
     });
   };
@@ -64,14 +91,27 @@ const NewApp: NextPage = () => {
         <ArrowLeftIcon aria-hidden className='h-4 w-4' />
         <span>{t('back')}</span>
       </Link>
-      <h2 className='mb-5 mt-5 font-bold text-gray-700 md:text-xl'>Add SAML Federation App</h2>
+      <h2 className='mb-5 mt-5 font-bold text-gray-700 md:text-xl'>Update SAML Federation App</h2>
       <div className='rounded border border-gray-200 bg-white p-6 dark:border-gray-700 dark:bg-gray-800'>
         <form onSubmit={onSubmit}>
           <div className='flex flex-col space-y-3'>
-            <p className='text-sm leading-6 text-gray-800'>
-              To configure SAML Federation app, add service provider details such as ACS URL and Entity ID.
-              You can find the details from your service provider portal or from the documentation.
-            </p>
+            <div className='form-control w-full md:w-1/2'>
+              <label className='label'>
+                <span className='label-text'>Tenant</span>
+              </label>
+              <input type='text' className='input-bordered input w-full' defaultValue={app.tenant} disabled />
+            </div>
+            <div className='form-control w-full md:w-1/2'>
+              <label className='label'>
+                <span className='label-text'>Product</span>
+              </label>
+              <input
+                type='text'
+                className='input-bordered input w-full'
+                defaultValue={app.product}
+                disabled
+              />
+            </div>
             <div className='form-control w-full md:w-1/2'>
               <label className='label'>
                 <span className='label-text'>App Name</span>
@@ -82,30 +122,7 @@ const NewApp: NextPage = () => {
                 className='input-bordered input w-full'
                 required
                 onChange={onChange}
-              />
-            </div>
-            <div className='form-control w-full md:w-1/2'>
-              <label className='label'>
-                <span className='label-text'>Tenant</span>
-              </label>
-              <input
-                type='text'
-                id='tenant'
-                className='input-bordered input w-full'
-                required
-                onChange={onChange}
-              />
-            </div>
-            <div className='form-control w-full md:w-1/2'>
-              <label className='label'>
-                <span className='label-text'>Product</span>
-              </label>
-              <input
-                type='text'
-                id='product'
-                className='input-bordered input w-full'
-                required
-                onChange={onChange}
+                value={existingApp.name}
               />
             </div>
             <div className='form-control w-full md:w-1/2'>
@@ -118,6 +135,7 @@ const NewApp: NextPage = () => {
                 className='input-bordered input w-full'
                 required
                 onChange={onChange}
+                value={existingApp.acsUrl}
               />
             </div>
             <div className='form-control w-full md:w-1/2'>
@@ -130,10 +148,13 @@ const NewApp: NextPage = () => {
                 className='input-bordered input w-full'
                 required
                 onChange={onChange}
+                value={existingApp.entityId}
               />
             </div>
             <div>
-              <button className={classNames('btn-primary btn', loading ? 'loading' : '')}>Create App</button>
+              <button className={classNames('btn-primary btn', loading ? 'loading' : '')}>
+                Save Changes
+              </button>
             </div>
           </div>
         </form>
@@ -142,12 +163,4 @@ const NewApp: NextPage = () => {
   );
 };
 
-// export const getServerSideProps: GetServerSideProps = async ({ locale }) => {
-//   return {
-//     props: {
-//       ...(await serverSideTranslations(locale ?? '', ['common'])),
-//     },
-//   };
-// };
-
-export default NewApp;
+export default UpdateApp;
